+++ cpp/src/link_prediction/jaccard.cu
@@ -19,7 +19,7 @@
  * @file jaccard.cu
  * ---------------------------------------------------------------------------**/
 
-#include "graph_utils.cuh"
+#include "utilities/graph_utils.cuh"
 #include "cugraph.h"
 #include "rmm_utils.h"
 #include "utilities/error_utils.h"
@@ -28,7 +28,7 @@ namespace cugraph {
   // Volume of neighboors (*weight_s)
   template<bool weighted, typename IdxType, typename ValType>
   __global__ void __launch_bounds__(CUDA_MAX_KERNEL_THREADS)
-  overlap_row_sum(IdxType n,
+  jaccard_row_sum(IdxType n,
                   IdxType *csrPtr,
                   IdxType *csrInd,
                   ValType *v,
@@ -56,7 +56,7 @@ namespace cugraph {
   // Volume of intersections (*weight_i) and cumulated volume of neighboors (*weight_s)
   template<bool weighted, typename IdxType, typename ValType>
   __global__ void __launch_bounds__(CUDA_MAX_KERNEL_THREADS)
-  overlap_is(IdxType n,
+  jaccard_is(IdxType n,
              IdxType *csrPtr,
              IdxType *csrInd,
              ValType *v,
@@ -81,7 +81,7 @@ namespace cugraph {
         cur = (Ni < Nj) ? col : row;
 
         //compute new sum weights
-        weight_s[j] = min(work[row], work[col]);
+        weight_s[j] = work[row] + work[col];
 
         //compute new intersection weights
         //search for the element with the same column index in the reference row
@@ -127,7 +127,7 @@ namespace cugraph {
   // Using list of node pairs
   template<bool weighted, typename IdxType, typename ValType>
   __global__ void __launch_bounds__(CUDA_MAX_KERNEL_THREADS)
-  overlap_is_pairs(IdxType num_pairs,
+  jaccard_is_pairs(IdxType num_pairs,
                    IdxType *csrPtr,
                    IdxType *csrInd,
                    IdxType *first_pair,
@@ -152,7 +152,7 @@ namespace cugraph {
       cur = (Ni < Nj) ? col : row;
 
       //compute new sum weights
-      weight_s[idx] = min(work[row], work[col]);
+      weight_s[idx] = work[row] + work[col];
 
       //compute new intersection weights
       //search for the element with the same column index in the reference row
@@ -197,26 +197,27 @@ namespace cugraph {
   //Jaccard  weights (*weight)
   template<bool weighted, typename IdxType, typename ValType>
   __global__ void __launch_bounds__(CUDA_MAX_KERNEL_THREADS)
-  overlap_jw(IdxType e,
+  jaccard_jw(IdxType e,
              IdxType *csrPtr,
              IdxType *csrInd,
              ValType *weight_i,
              ValType *weight_s,
              ValType *weight_j) {
     IdxType j;
-    ValType Wi, Wu;
+    ValType Wi, Ws, Wu;
 
     for (j = threadIdx.x + blockIdx.x * blockDim.x;
         j < e;
         j += gridDim.x * blockDim.x) {
       Wi = weight_i[j];
-      Wu = weight_s[j];
+      Ws = weight_s[j];
+      Wu = Ws - Wi;
       weight_j[j] = (Wi / Wu);
     }
   }
 
   template<bool weighted, typename IdxType, typename ValType>
-  int overlap(IdxType n,
+  int jaccard(IdxType n,
               IdxType e,
               IdxType *csrPtr,
               IdxType *csrInd,
@@ -236,7 +237,7 @@ namespace cugraph {
     nblocks.y = min((n + nthreads.y - 1) / nthreads.y, (IdxType) CUDA_MAX_BLOCKS);
     nblocks.z = 1;
     //launch kernel
-    overlap_row_sum<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(n,
+    jaccard_row_sum<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(n,
                                                                         csrPtr,
                                                                         csrInd,
                                                                         weight_in,
@@ -251,7 +252,7 @@ namespace cugraph {
     nblocks.y = 1;
     nblocks.z = min((n + nthreads.z - 1) / nthreads.z, (IdxType) CUDA_MAX_BLOCKS); //1;
     //launch kernel
-    overlap_is<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(n,
+    jaccard_is<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(n,
                                                                    csrPtr,
                                                                    csrInd,
                                                                    weight_in,
@@ -267,7 +268,7 @@ namespace cugraph {
     nblocks.y = 1;
     nblocks.z = 1;
     //launch kernel
-    overlap_jw<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(e,
+    jaccard_jw<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(e,
                                                                    csrPtr,
                                                                    csrInd,
                                                                    weight_i,
@@ -278,7 +279,7 @@ namespace cugraph {
   }
 
   template<bool weighted, typename IdxType, typename ValType>
-  int overlap_pairs(IdxType n,
+  int jaccard_pairs(IdxType n,
                     IdxType num_pairs,
                     IdxType *csrPtr,
                     IdxType *csrInd,
@@ -300,7 +301,7 @@ namespace cugraph {
     nblocks.y = min((n + nthreads.y - 1) / nthreads.y, (IdxType) CUDA_MAX_BLOCKS);
     nblocks.z = 1;
     //launch kernel
-    overlap_row_sum<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(n,
+    jaccard_row_sum<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(n,
                                                                         csrPtr,
                                                                         csrInd,
                                                                         weight_in,
@@ -315,7 +316,7 @@ namespace cugraph {
     nblocks.y = 1;
     nblocks.z = min((n + nthreads.z - 1) / nthreads.z, (IdxType) CUDA_MAX_BLOCKS); //1;
     //launch kernel
-    overlap_is_pairs<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(num_pairs,
+    jaccard_is_pairs<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(num_pairs,
                                                                          csrPtr,
                                                                          csrInd,
                                                                          first_pair,
@@ -333,7 +334,7 @@ namespace cugraph {
     nblocks.y = 1;
     nblocks.z = 1;
     //launch kernel
-    overlap_jw<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(num_pairs,
+    jaccard_jw<weighted, IdxType, ValType> <<<nblocks, nthreads>>>(num_pairs,
                                                                    csrPtr,
                                                                    csrInd,
                                                                    weight_i,
@@ -344,7 +345,7 @@ namespace cugraph {
   }
 } // End cugraph namespace
 
-gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result) {
+gdf_error gdf_jaccard(gdf_graph *graph, gdf_column *weights, gdf_column *result) {
   GDF_REQUIRE(graph != nullptr, GDF_INVALID_API_CALL);
   GDF_REQUIRE((graph->adjList != nullptr) || (graph->edgeList != nullptr), GDF_INVALID_API_CALL);
   GDF_REQUIRE(result != nullptr, GDF_INVALID_API_CALL);
@@ -375,7 +376,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
     ALLOC_TRY(&weight_i, sizeof(float) * e, nullptr);
     ALLOC_TRY(&weight_s, sizeof(float) * e, nullptr);
     ALLOC_TRY(&work, sizeof(float) * n, nullptr);
-    cugraph::overlap<true, int32_t, float>(n,
+    cugraph::jaccard<true, int32_t, float>(n,
                                            e,
                                            (int32_t*) csrPtr,
                                            (int32_t*) csrInd,
@@ -391,7 +392,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
     ALLOC_TRY(&weight_i, sizeof(float) * e, nullptr);
     ALLOC_TRY(&weight_s, sizeof(float) * e, nullptr);
     ALLOC_TRY(&work, sizeof(float) * n, nullptr);
-    cugraph::overlap<false, int32_t, float>(n,
+    cugraph::jaccard<false, int32_t, float>(n,
                                             e,
                                             (int32_t*) csrPtr,
                                             (int32_t*) csrInd,
@@ -407,7 +408,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
     ALLOC_TRY(&weight_i, sizeof(double) * e, nullptr);
     ALLOC_TRY(&weight_s, sizeof(double) * e, nullptr);
     ALLOC_TRY(&work, sizeof(double) * n, nullptr);
-    cugraph::overlap<true, int32_t, double>(n,
+    cugraph::jaccard<true, int32_t, double>(n,
                                             e,
                                             (int32_t*) csrPtr,
                                             (int32_t*) csrInd,
@@ -423,7 +424,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
     ALLOC_TRY(&weight_i, sizeof(double) * e, nullptr);
     ALLOC_TRY(&weight_s, sizeof(double) * e, nullptr);
     ALLOC_TRY(&work, sizeof(double) * n, nullptr);
-    cugraph::overlap<false, int32_t, double>(n,
+    cugraph::jaccard<false, int32_t, double>(n,
                                              e,
                                              (int32_t*) csrPtr,
                                              (int32_t*) csrInd,
@@ -439,7 +440,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
     ALLOC_TRY(&weight_i, sizeof(float) * e, nullptr);
     ALLOC_TRY(&weight_s, sizeof(float) * e, nullptr);
     ALLOC_TRY(&work, sizeof(float) * n, nullptr);
-    cugraph::overlap<true, int64_t, float>(n,
+    cugraph::jaccard<true, int64_t, float>(n,
                                            e,
                                            (int64_t*) csrPtr,
                                            (int64_t*) csrInd,
@@ -455,7 +456,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
     ALLOC_TRY(&weight_i, sizeof(float) * e, nullptr);
     ALLOC_TRY(&weight_s, sizeof(float) * e, nullptr);
     ALLOC_TRY(&work, sizeof(float) * n, nullptr);
-    cugraph::overlap<false, int64_t, float>(n,
+    cugraph::jaccard<false, int64_t, float>(n,
                                             e,
                                             (int64_t*) csrPtr,
                                             (int64_t*) csrInd,
@@ -471,7 +472,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
     ALLOC_TRY(&weight_i, sizeof(double) * e, nullptr);
     ALLOC_TRY(&weight_s, sizeof(double) * e, nullptr);
     ALLOC_TRY(&work, sizeof(double) * n, nullptr);
-    cugraph::overlap<true, int64_t, double>(n,
+    cugraph::jaccard<true, int64_t, double>(n,
                                             e,
                                             (int64_t*) csrPtr,
                                             (int64_t*) csrInd,
@@ -487,7 +488,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
     ALLOC_TRY(&weight_i, sizeof(double) * e, nullptr);
     ALLOC_TRY(&weight_s, sizeof(double) * e, nullptr);
     ALLOC_TRY(&work, sizeof(double) * n, nullptr);
-    cugraph::overlap<false, int64_t, double>(n,
+    cugraph::jaccard<false, int64_t, double>(n,
                                              e,
                                              (int64_t*) csrPtr,
                                              (int64_t*) csrInd,
@@ -506,7 +507,7 @@ gdf_error gdf_overlap(gdf_graph *graph, gdf_column *weights, gdf_column *result)
   return GDF_SUCCESS;
 }
 
-gdf_error gdf_overlap_list(gdf_graph* graph,
+gdf_error gdf_jaccard_list(gdf_graph* graph,
                            gdf_column* weights,
                            gdf_column* first,
                            gdf_column* second,
@@ -553,7 +554,7 @@ gdf_error gdf_overlap_list(gdf_graph* graph,
     ALLOC_TRY(&weight_i, sizeof(float) * num_pairs, nullptr);
     ALLOC_TRY(&weight_s, sizeof(float) * num_pairs, nullptr);
     ALLOC_TRY(&work, sizeof(float) * n, nullptr);
-    cugraph::overlap_pairs<true, int32_t, float>(n,
+    cugraph::jaccard_pairs<true, int32_t, float>(n,
                                                  num_pairs,
                                                  (int32_t*) csrPtr,
                                                  (int32_t*) csrInd,
@@ -572,7 +573,7 @@ gdf_error gdf_overlap_list(gdf_graph* graph,
     ALLOC_TRY(&weight_i, sizeof(float) * num_pairs, nullptr);
     ALLOC_TRY(&weight_s, sizeof(float) * num_pairs, nullptr);
     ALLOC_TRY(&work, sizeof(float) * n, nullptr);
-    cugraph::overlap_pairs<false, int32_t, float>(n,
+    cugraph::jaccard_pairs<false, int32_t, float>(n,
                                                   num_pairs,
                                                   (int32_t*) csrPtr,
                                                   (int32_t*) csrInd,
@@ -591,7 +592,7 @@ gdf_error gdf_overlap_list(gdf_graph* graph,
     ALLOC_TRY(&weight_i, sizeof(double) * num_pairs, nullptr);
     ALLOC_TRY(&weight_s, sizeof(double) * num_pairs, nullptr);
     ALLOC_TRY(&work, sizeof(double) * n, nullptr);
-    cugraph::overlap_pairs<true, int32_t, double>(n,
+    cugraph::jaccard_pairs<true, int32_t, double>(n,
                                                   num_pairs,
                                                   (int32_t*) csrPtr,
                                                   (int32_t*) csrInd,
@@ -610,7 +611,7 @@ gdf_error gdf_overlap_list(gdf_graph* graph,
     ALLOC_TRY(&weight_i, sizeof(double) * num_pairs, nullptr);
     ALLOC_TRY(&weight_s, sizeof(double) * num_pairs, nullptr);
     ALLOC_TRY(&work, sizeof(double) * n, nullptr);
-    cugraph::overlap_pairs<false, int32_t, double>(n,
+    cugraph::jaccard_pairs<false, int32_t, double>(n,
                                                    num_pairs,
                                                    (int32_t*) csrPtr,
                                                    (int32_t*) csrInd,
@@ -629,7 +630,7 @@ gdf_error gdf_overlap_list(gdf_graph* graph,
     ALLOC_TRY(&weight_i, sizeof(float) * num_pairs, nullptr);
     ALLOC_TRY(&weight_s, sizeof(float) * num_pairs, nullptr);
     ALLOC_TRY(&work, sizeof(float) * n, nullptr);
-    cugraph::overlap_pairs<true, int64_t, float>(n,
+    cugraph::jaccard_pairs<true, int64_t, float>(n,
                                                  num_pairs,
                                                  (int64_t*) csrPtr,
                                                  (int64_t*) csrInd,
@@ -648,7 +649,7 @@ gdf_error gdf_overlap_list(gdf_graph* graph,
     ALLOC_TRY(&weight_i, sizeof(float) * num_pairs, nullptr);
     ALLOC_TRY(&weight_s, sizeof(float) * num_pairs, nullptr);
     ALLOC_TRY(&work, sizeof(float) * n, nullptr);
-    cugraph::overlap_pairs<false, int64_t, float>(n,
+    cugraph::jaccard_pairs<false, int64_t, float>(n,
                                                   num_pairs,
                                                   (int64_t*) csrPtr,
                                                   (int64_t*) csrInd,
@@ -667,7 +668,7 @@ gdf_error gdf_overlap_list(gdf_graph* graph,
     ALLOC_TRY(&weight_i, sizeof(double) * num_pairs, nullptr);
     ALLOC_TRY(&weight_s, sizeof(double) * num_pairs, nullptr);
     ALLOC_TRY(&work, sizeof(double) * n, nullptr);
-    cugraph::overlap_pairs<true, int64_t, double>(n,
+    cugraph::jaccard_pairs<true, int64_t, double>(n,
                                                   num_pairs,
                                                   (int64_t*) csrPtr,
                                                   (int64_t*) csrInd,
@@ -686,7 +687,7 @@ gdf_error gdf_overlap_list(gdf_graph* graph,
     ALLOC_TRY(&weight_i, sizeof(double) * num_pairs, nullptr);
     ALLOC_TRY(&weight_s, sizeof(double) * num_pairs, nullptr);
     ALLOC_TRY(&work, sizeof(double) * n, nullptr);
-    cugraph::overlap_pairs<false, int64_t, double>(n,
+    cugraph::jaccard_pairs<false, int64_t, double>(n,
                                                    num_pairs,
                                                    (int64_t*) csrPtr,
                                                    (int64_t*) csrInd,